requirements.txt
pandas
numpy
scikit-learn

black

ruff

isort

mypy

pre-commit

pyyaml

types-PyYAML

mlflow
